Move Gear classes to their own package
