+++ src/main/kotlin/profile/gearset/Glamour.kt
@@ -1,3 +1,3 @@
-package cloud.drakon.ktlodestone.profile.gear
+package cloud.drakon.ktlodestone.profile.gearset
 
 data class Glamour(val name: String, val dbLink: String)
